Improve configuration handling Simplify meson build

+++ build.bat
@@ -3,43 +3,51 @@
 
 set CC=cl
 set TYPE=debug
+set ARCH=x64
 set CLEAN=
 
-if exist build-tmp\_clang set CC=clang
-if exist build-tmp\_release set TYPE=release
+set CONFIG=
+if exist build-tmp\_config set /p CONFIG=<build-tmp\_config
 
-if "%1"=="cl" set CC=cl&& shift /1
-if "%1"=="clang" set CC=clang&& shift /1
-if "%1"=="debug" set TYPE=debug&& shift /1
-if "%1"=="release" set TYPE=release&& shift /1
-if "%1"=="clean" set CLEAN=y&& shift /1
+for %%i in (%CONFIG% %*) do (
+	if "%%i"=="cl" set CC=cl
+	if "%%i"=="clang" set CC=clang
+	if "%%i"=="debug" set TYPE=debug
+	if "%%i"=="release" set TYPE=release
+	if "%%i"=="x64" set ARCH=x64
+	if "%%i"=="clean" set CLEAN=clean
+	if "%%i"=="rebuild" set CLEAN=y
+)
 
-if not exist build-tmp\_%CC% set CLEAN=y
-if not exist build-tmp\_%TYPE% set CLEAN=y
+if not "%CLEAN%"=="clean" if not "%CONFIG%"=="%CC% %ARCH% %TYPE%" (
+	echo Config changed from '%CONFIG%' to '%CC% %ARCH% %TYPE%'
+	set CLEAN=y
+)
 
 set vswhere=%ProgramFiles(x86)%\Microsoft Visual Studio\Installer\vswhere.exe
 if "%CC%%VCINSTALLDIR%"=="cl" if exist "%vswhere%" for /f "tokens=*" %%i in ('"%vswhere%" -latest -find VC') do (
 	set vcvars=%%i\Auxiliary\Build\vcvarsall.bat
 )
-if not "%vcvars%"=="" call "%vcvars%" x64
+if not "%vcvars%"=="" call "%vcvars%" %ARCH%
 
 if "%CC%%VCINSTALLDIR%"=="cl" (
-	echo Needs to run from a Visual Studio Developer Command Prompt
+	echo Requires a Visual Studio Developer Command Prompt
 	goto :eof
 )
 
 if not "%CLEAN%"=="" (
 	echo Cleaning...
-	rmdir /s /q build >nul:
-	rmdir /s /q build-tmp >nul:
+	if exist build rmdir /s /q build >nul:
+	if exist build-tmp rmdir /s /q build-tmp >nul:
 )
+
+if "%CLEAN%"=="clean" goto :eof
 
 echo Building %TYPE% with %CC%...
 if not exist build-tmp meson --buildtype=%TYPE% build-tmp
-echo > build-tmp\_%CC%
-echo > build-tmp\_%TYPE%
+echo %CC% %ARCH% %TYPE%>build-tmp\_config
 
 set DESTDIR=%CD%\build
 ninja -C build-tmp install
 
-if not "%vcvars%"=="" echo Built using "%vcvars%" x64
+if not "%vcvars%"=="" echo Built using "%vcvars%" %ARCH%
